Help Center: move useFormTitle to hooks
